fix(calling): update calling screen to fit design specs
