+++ app/src/main/kotlin/com/wire/android/ui/calling/initiating/InitiatingCallScreen.kt
@@ -1,6 +1,7 @@
 package com.wire.android.ui.calling.initiating
 
 import android.view.View
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,7 +12,6 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.BottomSheetScaffold
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.rememberBottomSheetScaffoldState
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -19,16 +19,15 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.wire.android.ui.calling.CallPreview
+import com.wire.android.ui.calling.common.CallVideoPreview
 import com.wire.android.ui.calling.CallState
 import com.wire.android.ui.calling.SharedCallingViewModel
+import com.wire.android.ui.calling.common.CallerDetails
 import com.wire.android.ui.calling.controlButtons.CallOptionsControls
 import com.wire.android.ui.calling.controlButtons.HangUpButton
 import com.wire.android.ui.common.dimensions
-import com.wire.android.ui.common.topappbar.WireCenterAlignedTopAppBar
 import com.wire.android.ui.theme.wireColorScheme
 import com.wire.android.ui.theme.wireDimensions
-import com.wire.android.util.EMPTY
 
 @Composable
 fun InitiatingCallScreen(
@@ -63,7 +62,6 @@ private fun InitiatingCallContent(
     val scaffoldState = rememberBottomSheetScaffoldState()
 
     BottomSheetScaffold(
-        topBar = { InitiatingCallTopBar { onNavigateBack() } },
         sheetShape = RoundedCornerShape(topStart = dimensions().corner16x, topEnd = dimensions().corner16x),
         backgroundColor = MaterialTheme.wireColorScheme.initiatingCallBackground,
         scaffoldState = scaffoldState,
@@ -96,28 +94,19 @@ private fun InitiatingCallContent(
             }
         }
     ) {
-        CallPreview(
-            conversationName = callState.conversationName,
-            isCameraOn = callState.isCameraOn,
-            avatarAssetId = callState.avatarAssetId,
-            conversationType = callState.conversationType,
-            onVideoPreviewCreated = { onVideoPreviewCreated(it) }
-        )
+        Box {
+            CallVideoPreview(
+                isCameraOn = callState.isCameraOn,
+                onVideoPreviewCreated = { onVideoPreviewCreated(it) }
+            )
+            CallerDetails(
+                conversationName = callState.conversationName,
+                isCameraOn = callState.isCameraOn,
+                avatarAssetId = callState.avatarAssetId,
+                conversationType = callState.conversationType,
+            )
+        }
     }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun InitiatingCallTopBar(
-    onNavigateBack: () -> Unit
-) {
-    WireCenterAlignedTopAppBar(
-        onNavigationPressed = onNavigateBack,
-        title = String.EMPTY,
-        navigationIconType = null,
-        elevation = 0.dp,
-        actions = { }
-    )
 }
 
 @Preview
